fix: correct tile visibility, physics, audio, and type safety bugs
- Fix off-by-one in tile visibility calculation (Console.h) by removing the extra `+ tileW - 1` term, ensuring the last visible tile is correctly computed.
- Guard against division by zero in `lerpi` (GameUtils.h) by returning `b` when `tmax <= 0`.
- Fix AABB sweep collision for zero-velocity axes (PhysicsEngine.h): when velocity is zero, set both entry and exit times to `-inf` if boxes are already separated, preventing false collision detection.
- Initialize all union members in `CacheData` constructors (SaveManager.H) by zeroing `u` before assigning the specific field, avoiding undefined behavior from reading uninitialized bytes.
- Replace `random()` calls in audio synthesis (SynthEngine.cpp) with a fast, thread-safe LCG noise generator to avoid contention and non-reentrant issues.
- Guard envelope calculations against zero attack/release times to prevent division by zero and ensure correct envelope shape when duration is less than release time.
- Add generated CMake build cache files for the simulator project.

+++ lib/Core/Diagnostics.h
@@ -72,14 +72,14 @@ public:
         uint32_t maxBlockKb = maxBlock / 1024;
         
         ctx.drawPrintf(2, 6, "%uF | %lu.%lums | R:%luK | M:%luK", 
-                       _fps, 
-                       logicTimeUs / 1000, (logicTimeUs % 1000) / 100, 
-                       freeHeapKb, 
-                       maxBlockKb);
+                       (unsigned int)_fps, 
+                       (unsigned long)(logicTimeUs / 1000), (unsigned long)((logicTimeUs % 1000) / 100), 
+                       (unsigned long)freeHeapKb, 
+                       (unsigned long)maxBlockKb);
 #else
         ctx.drawPrintf(2, 6, "%uF | %lu.%lums | PC Sim", 
-                       _fps, 
-                       logicTimeUs / 1000, (logicTimeUs % 1000) / 100);
+                       (unsigned int)_fps, 
+                       (unsigned long)(logicTimeUs / 1000), (unsigned long)((logicTimeUs % 1000) / 100));
 #endif
         ctx.endScreenSpace();
     }
